Add files via upload
Lecture 11 is about how to apply newton method to problem with inequality constraints. Here, we use indicator functions to remove inequality constraints and introduce a log barrier function to approximate the indicator function. Then, we do some transformation to get the range of optimal value p*. The process to get p* consists of two nested loops that are called the outer iteration and inner iteration in barrier method. Finally, we use Phase I to check whether there is a strictly feasible solution to the original problem.